Remove all propertyChanged signal arguments
Signal arguments will be added into the scope of the
receiving QML functions/slots as undeclared variables.
This is problematic, since it can silently clash and
shadow the applications own local variables. This becomes
even more problematic when the variable name is the same
as the property that changed, since then you will
always shadow the property, which is confusing and error
prone.

This patch will remove all signal changed arguments in
the front end QML classes.

Fixes: QTBUG-79575
Change-Id: Id45c1bf2ab4c6d33ce93c9681e1ffad3191ccf1d