QQuick3DNode: change sceneRotation implementation
It turns out that you cannot extract the scene rotation
directly out of the scene matrix if the node has an ancestor
that is non-uniformly scaled. If the ancestor has a non-uniform
scale and a rotation, it can end up skewed, which means that the
axis are no longer perpendicular.

So we need to detect this case, and fall back to calculate the
sceneRotation by explicitly multiplying the rotation matrices
without including the scale. Since this path is less performant
(if we assume that we anyway need to calculate the scene matrix)
we detect and store if we have a non-uniformly scaled ancestor, and
choose implementation strategy based on that.

Change-Id: I4fef98ad08ae8e646361807066899a0fb8798877